# backend/cmd/app/docs/swagger.yaml
basePath: /api
definitions:
...
  auth.EmailRequest:
    properties:
      email:
        example: user@example.com
        type: string
    type: object
  auth.ResetRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: newPassword123
        type: string
      token:
        example: abc123
        type: string
    type: object
...
  data.Error:
    properties:
      error:
        type: string
    type: object
...
  data.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  data.MerchantSignupRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      nameOfStore:
        type: string
      password:
        type: string
      phone:
        type: string
      storeDescription:
        type: string
    type: object
  data.OtpRequest:
    properties:
      code:
        type: string
      email:
        type: string
    type: object
  data.RiderSignupRequest:
    properties:
      confirmPassword:
        type: string
      deliveryServiceCode:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  data.SignupRequest:
    properties:
      confirmPassword:
        type: string
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
  description: Boiboi API Documentation.
  title: Boiboi Backend API
  version: "1.0"
paths:
...
  /auth/forgotPassword:
    post:
      consumes:
      - application/json
      description: Send password reset link to user's email address
      parameters:
      - description: Email address for password reset
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid email format
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error - user not found or email sending failed
          schema:
            $ref: '#/definitions/data.Error'
      summary: Request password reset
      tags:
      - Authentication
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/data.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with JWT token and user data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid credentials or rider not approved
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/data.Error'
      summary: User login
      tags:
      - Authentication
  /auth/merchantSignup:
    post:
      consumes:
      - application/json
      description: Register a new merchant account with store information and email
        verification via OTP
      parameters:
      - description: Merchant registration details including store information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/data.MerchantSignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error or email already exists
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/data.Error'
      summary: Register a new merchant account
      tags:
      - Authentication
  /auth/resetPassword:
    post:
      consumes:
      - application/json
      description: Reset user password using email and reset token
      parameters:
      - description: Password reset details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.ResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid request format
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error - invalid token or user not found
          schema:
            $ref: '#/definitions/data.Error'
      summary: Reset user password
      tags:
      - Authentication
  /auth/riderSignup:
    post:
      consumes:
      - application/json
      description: Register a new rider account with delivery service code and email
        verification via OTP
      parameters:
      - description: Rider registration details including delivery service code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/data.RiderSignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error or email already exists
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/data.Error'
      summary: Register a new rider account
      tags:
      - Authentication
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new base user account with email verification via OTP
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/data.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - validation error or email already exists
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/data.Error'
      summary: Register a new user account
      tags:
      - Authentication
  /auth/verifyMerchantSignup:
    post:
      consumes:
      - application/json
      description: Verify the OTP code sent to merchant's email and complete the registration
        process including store creation
      parameters:
      - description: OTP verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/data.OtpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Merchant and store created successfully with JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid OTP or validation error
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/data.Error'
      summary: Verify merchant signup with OTP
      tags:
      - Authentication
  /auth/verifyRiderSignup:
    post:
      consumes:
      - application/json
      description: Verify the OTP code sent to rider's email and complete the registration
        process
      parameters:
      - description: OTP verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/data.OtpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rider created successfully with JWT token or pending status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid OTP or validation error
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/data.Error'
      summary: Verify rider signup with OTP
      tags:
      - Authentication
  /auth/verifySignup:
    post:
      consumes:
      - application/json
      description: Verify the OTP code sent to user's email and complete the registration
        process
      parameters:
      - description: OTP verification details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/data.OtpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully with JWT token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request - invalid OTP or validation error
          schema:
            $ref: '#/definitions/data.Error'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/data.Error'
      summary: Verify user signup with OTP
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"